+++ LootMethods/WhitelistLoot.cs
@@ -38,7 +38,7 @@ namespace LootManager
             // FilterList check (all active categories)
             foreach (var kvp in Plugin.FilterList)
             {
-                if (!Plugin.EnabledFilterCategories.Contains(kvp.Key))
+                if (!Plugin.EnabledFilterGroups.Contains(kvp.Key))
                     continue; // Skip if not enabled
 
                 if (kvp.Value.Contains(item.ItemName))